--- src/VideoPage.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { useParams } from 'react-router';
-
-export function VideoPage() {
-  const { videoID } = useParams();
-  return (
-    <div>
-      <iframe width="560" height="315" src={`https://www.youtube.com/embed/${videoID}`} title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowFullscreen="allowfullscreen"></iframe>
-    </div>
-  );
-}
