+++ src/components/HeaderMobile/HeaderMobile.js
@@ -1,0 +1,8 @@
+/*
+ * --------------TODO--------------
+ * Header mobile goes at top including logo, search, notifications and user profile
+ */
+
+export function HeaderMobile() {
+  return <h1>Header mobile will come here</h1>;
+}
