+++ src/components/NavbarMobile/NavbarMobile.js
@@ -1,0 +1,9 @@
+/*
+ * --------------TODO--------------
+ * Navbar mobile to stick at bottom with Home, trending and library
+ * (which will consist of sign in if not signed in or history, watch later, liked videos and playlists as lists)
+ */
+
+export function NavbarMobile() {
+  return <h1>Nav mobile will stick at bot </h1>;
+}
