+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { NavbarMobile } from './NavbarMobile/NavbarMobile';
+export { HeaderMobile } from './HeaderMobile/HeaderMobile';
