+++ src/context/index.js
@@ -1,0 +1,1 @@
+export { useData, DataProvider } from './dataContext';
