+++ src/pages/Error/Error404.js
@@ -1,3 +1,3 @@
-export function NoMatch() {
+export function Error404() {
   return <h1>Error 404</h1>;
 }
