+++ src/pages/Playlist/components/PlaylistCard.js
@@ -1,5 +1,5 @@
 import { PlaylistVideoCard } from './PlaylistVideoCard';
-import { useData } from './dataContext';
+import { useData } from '../../../context';
 import { useRef, useState } from 'react';
 
 /*
