+++ src/pages/WatchLater/components/WatchLaterCard.js
@@ -1,9 +1,8 @@
-import React from 'react';
-import { useData } from './dataContext';
+import { useData } from '../../../context';
 
-export function LikedVideosCard({ video }) {
+export function WatchLaterCard({ video }) {
   const { id, name, url, channel, thumbnail, views, time } = video;
-  const { videoDispatch, toggleWatchLaterText, toggleLikeButtonText } = useData();
+  const { videoDispatch, toggleWatchLaterText } = useData();
 
   return (
     <div
@@ -24,7 +23,6 @@ export function LikedVideosCard({ video }) {
       <div> {channel} </div>
       <div> {time} </div>
       <button onClick={() => videoDispatch({ type: 'TOGGLE_WATCH_LATER', payload: video })}>{toggleWatchLaterText(id)}</button>
-      <button onClick={() => videoDispatch({ type: 'TOGGLE_LIKED_VIDEO', payload: video })}>{toggleLikeButtonText(id)}</button>
     </div>
   );
 }
