+++ src/pages/index.js
@@ -1,0 +1,7 @@
+export { History } from './History/History';
+export { VideoListing } from './Home/VideoListing';
+export { VideoPage } from './Home/VideoPage';
+export { LikedVideos } from './LikedVideos/LikedVideos';
+export { Playlist } from './Playlist/Playlist';
+export { WatchLater } from './WatchLater/WatchLater';
+export { Error404 } from './Error/Error404';
